+++ app/src/main/java/com/nextcloud/client/di/AppComponent.java
@@ -22,7 +22,7 @@ package com.nextcloud.client.di;
 
 import android.app.Application;
 
-import com.nextcloud.app_review.InAppReviewModule;
+import com.nextcloud.appReview.InAppReviewModule;
 import com.nextcloud.client.appinfo.AppInfoModule;
 import com.nextcloud.client.database.DatabaseModule;
 import com.nextcloud.client.device.DeviceModule;
