+++ app/src/main/java/com/nextcloud/client/preferences/AppPreferencesImpl.java
@@ -26,7 +26,7 @@ import android.content.Context;
 import android.content.SharedPreferences;
 
 import com.google.gson.Gson;
-import com.nextcloud.app_review.AppReviewShownModel;
+import com.nextcloud.appReview.AppReviewShownModel;
 import com.nextcloud.client.account.User;
 import com.nextcloud.client.account.UserAccountManager;
 import com.nextcloud.client.account.UserAccountManagerImpl;
