+++ app/src/test/java/com/nextcloud/android/utils/ExtensionsTest.kt
